Backend: Added user validation for the reviews apis

+++ Backend/controllers/reviews.js
@@ -26,14 +26,19 @@ const addReview = async (req, res) => {
 
 const editReview = async (req, res) => {
     try {
-        const review = await Review.findById(req.params.id);
-        //Verify that the review belongs to the user then update
-        if (review.user_id.equals(req.body.user_id)){
-          await review.updateOne({ $set: {text: req.body.text}});
-          const edited_review = await Review.findById(req.params.id);
-          res.status(200).send({message:"Review successfully edited!", review: edited_review});
-        } else {
-          res.status(403).send("You are not allowed to edit others' reviews.");
+        const user_match = await User.findOne( {_id: req.body.user_id});
+        if (user_match){
+            const review = await Review.findById(req.params.id);
+            //Verify that the review belongs to the user then update
+            if (review.user_id.equals(req.body.user_id)){
+              await review.updateOne({ $set: {text: req.body.text}});
+              const edited_review = await Review.findById(req.params.id);
+              res.status(200).send({message:"Review successfully edited!", review: edited_review});
+            } else {
+              res.status(403).send("You are not allowed to edit others' reviews.");
+            }
+        }else{
+            res.status(401).send("Invalid user.");
         }
     } catch (err) {
         res.status(500).send(err);
@@ -42,15 +47,21 @@ const editReview = async (req, res) => {
 
 const deleteReview =  async (req, res) => {
     try {
-      const review = await Review.findById(req.params.id);
-      const user = await User.findById(review.user_id);
-      if (review.user_id.equals(req.body.user_id)) {
-        await review.deleteOne();
-        await user.updateOne({ $pull: { reviews: req.params.id } });
-        res.status(200).send("Review successfully deleted!");
-      } else {
-        res.status(403).send("You are not allowed to delete others' reviews.");
-      }
+        const user_match = await User.findOne( {_id: req.body.user_id});
+
+        if(user_match){
+            const review = await Review.findById(req.params.id);
+            const user = await User.findById(review.user_id);
+            if (review.user_id.equals(req.body.user_id)) {
+              await review.deleteOne();
+              await user.updateOne({ $pull: { reviews: req.params.id } });
+              res.status(200).send("Review successfully deleted!");
+            } else {
+              res.status(403).send("You are not allowed to delete others' reviews.");
+            }
+        } else {
+            res.status(401).send("Invalid user.");
+        }     
     } catch (err) {
       res.status(500).send(err);
     }
@@ -58,15 +69,21 @@ const deleteReview =  async (req, res) => {
 
 const likeReview = async (req, res) => {
     try {
-        const review = await Review.findById(req.params.id);
-        if (!review.likes.includes(req.body.user_id)) {
-          await review.updateOne({ $push: { likes: req.body.user_id } });
-          const liked_review = await Review.findById(req.params.id);
-          res.status(200).send({message:"Review liked successfully!", review:liked_review});
-        } else {
-          await review.updateOne({ $pull: { likes: req.body.user_id } });
-          const liked_review = await Review.findById(req.params.id);
-          res.status(200).send({message:"Like removed", review:liked_review});
+        const user_match = await User.findOne( {_id: req.body.user_id});
+
+        if(user_match){
+            const review = await Review.findById(req.params.id);
+            if (!review.likes.includes(req.body.user_id)) {
+              await review.updateOne({ $push: { likes: req.body.user_id } });
+              const liked_review = await Review.findById(req.params.id);
+              res.status(200).send({message:"Review liked successfully!", review:liked_review});
+            } else {
+              await review.updateOne({ $pull: { likes: req.body.user_id } });
+              const liked_review = await Review.findById(req.params.id);
+              res.status(200).send({message:"Like removed", review:liked_review});
+            }
+        }else{
+            res.status(401).send("Invalid user.");
         }
       } catch (err) {
         res.status(500).send(err);
@@ -76,14 +93,20 @@ const likeReview = async (req, res) => {
 const dislikeReview = async (req, res) => {
     try {
         const review = await Review.findById(req.params.id);
-        if (!review.dislikes.includes(req.body.user_id)) {
-          await review.updateOne({ $push: { dislikes: req.body.user_id } });
-          const disliked_review = await Review.findById(req.params.id);
-          res.status(200).send({message:"Review disliked successfully!", review:disliked_review});
-        } else {
-          await review.updateOne({ $pull: { dislikes: req.body.user_id } });
-          const disliked_review = await Review.findById(req.params.id);
-          res.status(200).send({message:"Dislike removed", review:disliked_review});
+        const user_match = await User.findOne( {_id: req.body.user_id});
+        
+        if(user_match){
+            if (!review.dislikes.includes(req.body.user_id)) {
+              await review.updateOne({ $push: { dislikes: req.body.user_id } });
+              const disliked_review = await Review.findById(req.params.id);
+              res.status(200).send({message:"Review disliked successfully!", review:disliked_review});
+            } else {
+              await review.updateOne({ $pull: { dislikes: req.body.user_id } });
+              const disliked_review = await Review.findById(req.params.id);
+              res.status(200).send({message:"Dislike removed", review:disliked_review});
+            }
+        }else{
+            res.status(401).send("Invalid user");
         }
       } catch (err) {
         res.status(500).send(err);
@@ -91,17 +114,25 @@ const dislikeReview = async (req, res) => {
 }
 
 const getUserReviews = async (req, res) => {
-
     try{
-        const reviews = await Review.find({user_id: req.body.user_id});
-        if (reviews.length) {
-            return res.status(200).send(reviews);
+        const user_match = await User.findOne({ _id: req.body.user_id });
+        if (user_match){
+            const reviews = await Review.find({user_id: req.body.user_id});
+            if (reviews.length) {
+                return res.status(200).send(reviews);
+            }else {
+                return res.status(400).send("No reviews");
+            }   
         }else {
-            return res.status(400).send("No reviews");
-        }   
-    } catch (error) {
+            return res.status(401).send("Invalid user")
+        }
+    }catch (err) {
         return res.status(500).send(err)   
     }
+}
+
+const getFeedReviews = async (req, res) => {
+
 }
 
 module.exports.addReview = addReview;
@@ -110,3 +141,4 @@ module.exports.deleteReview = deleteReview;
 module.exports.likeReview = likeReview;
 module.exports.dislikeReview = dislikeReview;
 module.exports.getUserReviews = getUserReviews;
+module.exports.getFeedReviews = getUserReviews;
